requirements.txt
bs4
requests
tqdm
selenium 
webdriver-manager
playwright
datasets 
huggingface_hub
Pillow
cairosvg
pandas
pytest